pyproject.toml
[project]
name = "customer-support-openenv"
version = "0.1.0"
description = "Customer Support Triage OpenEnv environment"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
  "fastapi",
  "uvicorn[standard]",
  "pydantic",
  "httpx",
  "openai",
  "python-dotenv",
  "openenv-core>=0.2.0"
]

[project.scripts]
server = "server.app:main"

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"